requirements.txt
Flask-Smorest
Flask-SQLAlchemy
Flask-Migrate
marshmallow
marshmallow-sqlalchemy
python-dotenv